% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Elim_rules.R
\name{get_Elim_rules}
\alias{get_Elim_rules}
\title{Get dose elimination rules for phase I design}
\usage{
get_Elim_rules(Nmax, phi, maxprob)
}
\arguments{
\item{Nmax}{maximum number of patients}

\item{phi}{target DLT rate}

\item{maxprob}{P(DLT rate>phi|c,n)<=maxprob}
}
\value{
a data.frame with elements
\itemize{
\item phi: target toxicity rate
\item n: number of patients at which toxicity is evaluated
\item c: max number of DLTs at which P(DLT rate>phi)<=maxprob
\item c_STOP: c+1, i.e. min number of DLTs at which P(DLT rate>phi|c_STOP,n)>maxprob
\item prob_Bayes: P(DLT rate>phi|c,n)
\item prob_Obs: c/n or the allowed observed toxicity rate with no stopping
}
}
\description{
Get Dose elimination rules for phase I design, based on Bayesian posterior probability
}
\details{
A uniform Beta(1,1) prior is used to calculate the posterior Bayesian probability
}
\examples{
rule<-get_Elim_rules(Nmax=21,phi=0.3,maxprob=0.95)
}
