% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fibonacci.R
\name{Get_Fibonacci}
\alias{Get_Fibonacci}
\title{Get 'modified Finonacci' sequence for Phase I design}
\usage{
Get_Fibonacci(
  Ndose,
  first_dose,
  fibonacci = c(1, 2, 1.67, 1.5, rep(1.33, Ndose - 4))
)
}
\arguments{
\item{Ndose}{number of dose levels}

\item{first_dose}{First dose level}

\item{fibonacci}{decreasing step scheme, default is c(1,2,1.67,1.5,rep(1.33,Ndose-4)}
}
\value{
a data.frame with elements
\itemize{
\item dose_level: dose level
\item dose: actual dose to be used
\item incr_ratio: The ratio of successive doses
}
}
\description{
Get 'modified Finonacci' sequence for Phase I design
    Other decreasing step schemes can be specified
}
\examples{
Fibonacci<-Get_Fibonacci(Ndose=6,first_dose=0.0017)
}
\references{
Penel N, Kramar A. What does a modified-Fibonacci dose-escalation
}
