% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_i33_rules.R
\name{get_i33_rules}
\alias{get_i33_rules}
\title{Get i3+3 decision rules for phase I design}
\usage{
get_i33_rules(N, phi, phi1, phi2)
}
\arguments{
\item{N}{sample size}

\item{phi}{target toxicity rate}

\item{phi1}{the highest DLT rate that is deemed subtherapeutic (i.e., underdosing), 
such that dose escalation should be made}

\item{phi2}{the lowest DLT rate that is deemed overly toxic (i.e., overdosing), 
such that dose de-escalation is required}
}
\value{
a data.frame with elements
\itemize{
\item n: number of patients
\item x: observed DLTs
\item decision: "E" (escalate),"R" (retain) or "D" (de-escalate)
}
}
\description{
Get i3+3 decision rules for phase I design
}
\examples{
rule<-get_i33_rules(N=6,phi=0.3,phi1=0.25,phi2=0.35)
}
\references{
Liu M et al. The i3+3 design for phase I clinical trials. 
    Journal of Biopharmaceutical Statistics 2020;30:294-304
}
