% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Keyboard_rules.R
\name{get_Keyboard_rules}
\alias{get_Keyboard_rules}
\title{Get Keyboard decision rules for phase I design}
\usage{
get_Keyboard_rules(N, phi, halfkey)
}
\arguments{
\item{N}{sample size}

\item{phi}{target toxicity rate}

\item{halfkey}{half length of key in keyboard design}
}
\value{
a data.frame with elements
\itemize{
\item n: number of patients
\item x: observed DLTs
\item decision: "E" (escalate),"R" (retain) or "D" (de-escalate)
}
}
\description{
Get Keyboard decision rules for phase I design
}
\details{
A uniform Beta(1,1) prior is used to calculate the posterior Bayesian probability for each 'key'
}
\examples{
Keyboard_rules <- get_Keyboard_rules(N=9,phi=0.3,halfkey=0.05)
}
\references{
Yan F, Mandrekar SJ, Yuan Y. Keyboard: A Novel Bayesian Toxicity Probability Interval Design for Phase I Clinical Trials. Clin Cancer Res. 2017; 23: 3994–4003
}
