% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Table_rules.R
\name{get_Table_rules}
\alias{get_Table_rules}
\title{Get table with summary of decision rules for phase I design}
\usage{
get_Table_rules(
  N,
  phi,
  phi1 = NULL,
  phi2 = NULL,
  maxtox,
  halfkey = NULL,
  design
)
}
\arguments{
\item{N}{sample size}

\item{phi}{target toxicity rate}

\item{phi1}{the highest DLT rate that is deemed subtherapeutic (i.e., underdosing), 
such that dose escalation should be made}

\item{phi2}{the lowest DLT rate that is deemed overly toxic (i.e., overdosing), 
such that dose de-escalation is required}

\item{maxtox}{P(DLT rate>phi|c,n)<=maxtox as value for maxprob in \code{\link{get_Elim_rules}}}

\item{halfkey}{half length of key in keyboard design}

\item{design}{"BOIN" or "Keyboard"}
}
\value{
a data.frame with elements
\itemize{
\item n: number of patients
\item E: max(observed DLTs) at n --> decision to escalate dose
\item R: range(observed DLTs) at n --> decision to retain dose
\item R: range(observed DLTs) at n --> decision to retain dose
\item lambda_e: if observed DLT rate<= lambda_e --> dose escalation
\item lambda_d: if observed DLT rate>= lambda_d --> dose de-escalation
}
}
\description{
Get table with summary of decision rules for phase I design (BOIN or Keyboard design)
}
\examples{
BOIN_table<-get_Table_rules(N=21,phi=0.3,phi1=0.6*0.3,phi2=1.4*0.3,maxtox=0.95,design="BOIN")
KEYB_table<-get_Table_rules(N=21,phi=0.3,maxtox=0.95,halfkey=0.05,design="Keyboard")
}
\references{
Liu S, Yuan Y. Bayesian Optimal Interval Designs for Phase I Clinical Trials. 
    J R Stat Soc Ser C Appl Stat. 2015;64:507–23
}
