% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph1_1sim.R
\name{ph1_1sim}
\alias{ph1_1sim}
\title{Simulation of one Phase I study}
\usage{
ph1_1sim(
  phi,
  phi1 = NULL,
  phi2 = NULL,
  maxtox = NULL,
  N = NULL,
  truerate = NULL,
  cohortsize = NULL,
  maxNretain = NULL,
  acc_tit,
  design,
  MTD_safer = NULL,
  seed = NULL,
  halfkey = NULL,
  sim = "NO",
  env = parent.frame()
)
}
\arguments{
\item{phi}{target toxicity rate (needed for all designs)}

\item{phi1}{for BOIN/i3+3 design: the highest DLT rate that is deemed subtherapeutic (i.e., underdosing), 
such that dose escalation should be made, see \code{\link{get_BOIN_rules}}}

\item{phi2}{for BOIN/i3+3 design: the lowest DLT rate that is deemed overly toxic (i.e., overdosing), 
such that dose de-escalation is required, see \code{\link{get_BOIN_rules}}}

\item{maxtox}{P(DLT rate>phi|c,n)<=maxtox as value for maxprob in \code{\link{get_Elim_rules}}}

\item{N}{number of patients (not needed for design="3+3")}

\item{truerate}{scenario of true DLT rates for each dose level. This also defines the number of dose levels}

\item{cohortsize}{cohort size, default is 3}

\item{maxNretain}{If N(patients) at next dose level >=maxNretain, then stop algorithm, default is 9}

\item{acc_tit}{logical indicator if algorithm should start with accelerated titration=algorithm starts at dose level with first DLT}

\item{design}{"BOIN" or "Keyboard" or "i3+3"}

\item{MTD_safer}{logical indicator for ...}

\item{seed}{define seed}

\item{halfkey}{for Keyboard design: half length of key in keyboard design, see \code{\link{get_Keyboard_rules}}}

\item{sim}{if "NO", calculate decision rules, define "YES" if done within simulation program for multiple studies}

\item{env}{parent environment to pass objects from \code{\link{ph1_sim_OC}}}
}
\value{
a data.frame with elements
\itemize{
\item dose_level: dose level
\item dose: actual dose to be used
\item incr_ratio: The ratio of successive doses
}
}
\description{
Simulation of one Phase I study with classical 3+3 design, 
    BOIN, Keyboard or i3+3 design
}
\examples{
ph1_1sim(phi=0.3, phi1=0.6*0.3, phi2=1.4*0.3, maxtox=0.95, N=15, 
truerate=c(0.20 ,0.25 ,0.30 ,0.40 ,0.60 ,0.75), cohortsize=3, maxNretain=9, acc_tit=0, 
design="BOIN", MTD_safer=TRUE)
ph1_1sim(phi=0.3, phi1=0.6*0.3, phi2=1.4*0.3, maxtox=0.95, N=30, 
truerate=c(0.20 ,0.25 ,0.30 ,0.40 ,0.60 ,0.75), cohortsize=3, maxNretain=9, acc_tit=1, 
design="Keyboard", MTD_safer=TRUE, halfkey=0.05)
ph1_1sim(phi=0.3, phi1=0.6*0.3, phi2=1.4*0.3, maxtox=0.95, N=15, 
truerate=c(0.20 ,0.25 ,0.30 ,0.40 ,0.60 ,0.75), cohortsize=3, maxNretain=9, acc_tit=0, 
design="i3+3", MTD_safer=TRUE)
ph1_1sim(truerate=c(0.20 ,0.25 ,0.30 ,0.40 ,0.60 ,0.75),acc_tit=1,design="3+3")
}
\references{
Liu S, Yuan Y. Bayesian Optimal Interval Designs for Phase I Clinical Trials. 
    J R Stat Soc Ser C Appl Stat. 2015;64:507–23
    Yan F, Mandrekar SJ, Yuan Y. Keyboard: A Novel Bayesian Toxicity Probability Interval Design for Phase I Clinical Trials. Clin Cancer Res. 2017; 23: 3994–4003
}
