% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph1_sim_OC.R
\name{ph1_sim_OC}
\alias{ph1_sim_OC}
\title{Simulation of Phase I study operating characteristics}
\usage{
ph1_sim_OC(
  nsim,
  scenarios,
  env = parent.frame(),
  OD_cut = c(0.6, 0.8),
  OD_delta = 0,
  ...
)
}
\arguments{
\item{nsim}{number of simulations}

\item{scenarios}{list of scenario's (each vector: set of true DLT rate for each dose level). 
Examples of scenarios can be found in a vignette accompanying this package. 
For more details see the help vignette: \code{vignette("scenarios", package = "PhIdesign")}}

\item{env}{parent environment (global) to pass scenario counter to global environment to print progress}

\item{OD_cut}{% of patients > OD_cut with overdose for overdose_1 and overdose_2 outcomes; should be 
vector of 2. Default value is c(0.6,0.8)}

\item{OD_delta}{dose level with toxicity >phi+OD_delta is considered as overdose. Default value is zero.}

\item{...}{see \code{\link{ph1_1sim}}}
}
\value{
a list containing of a matrix (results by dose level) and a vector (overall results)
\itemize{
\item [[1]]$dose_level: dose level
\item [[1]]$truerate: true DLT rate
\item [[1]]$sel_pct_dose: selection proportion at each dose level
\item [[1]]$avg_SS_dose: average number of patients treated by dose level
\item [[1]]$nDLT_dose: average number of DLT's observed at dose level
\item [[2]]$nsim: number of simulations
\item [[2]]$design: design used
\item [[2]]$sel_at_phi_pct: percent simulated trials in which the correct dose is selected, i.e., the dose with DLT rate equal to phi
\item [[2]]$sel_at_phi_rng_pct: percent simulated trials in which the selected dose has a DLT rate that lies in the interval [phi-0.05, phi+0.05]
\item [[2]]$sel_below_phi_pct: percent simulated trials in which the selected dose has a DLT rate < phi-0.05
\item [[2]]$sel_above_phi_pct: percent simulated trials in which the selected dose has a DLT rate > phi+0.05
\item [[2]]$avg_npt_MTD_sel: average number of patients treated at the actual selected MTD for each simulated study (studies without MTD selection not taken into account)
\item [[2]]$avg_pct_pts_at_phi: average percent patients in each simulated trial treated with dose with DLT rate =phi
\item [[2]]$avg_pct_pts_at_phi_rng: average percent patients in each simulated trial treated with dose with DLT rate that lies in the interval [phi-0.05, phi+0.05]
\item [[2]]$avg_pct_pts_below_phi: average percent patients in each simulated trial treated with dose with DLT rate < phi-0.05
\item [[2]]$avg_pct_pts_above_phi: average percent patients in the simulated trials treated with dose with DLT rate > phi+0.05
\item [[2]]$noMTD: percent simulated trials without MTD selection
\item [[2]]$avg_nDLT: average number of observed DLTs
\item [[2]]$avg_SS: average sample size of all simulated trials
\item [[2]]$max_SS: maximum sample size of all simulated trials
\item [[2]]$overdose_1: percent simulated trials in which >=OD_cut[1] of patients is assigned to a dose with DLT rate >phi
\item [[2]]$overdose_2: percent simulated trials in which >=OD_cut[2] of patients is assigned to a dose with DLT rate >phi
\item [[3]] labels for item[[2]]
}
}
\description{
Simulation of Phase I study operating characteristics
}
\examples{
result<-ph1_sim_OC(nsim=5,  scenarios=list(c(0.40 ,0.50 ,0.60 ,0.70 ,0.80 ,0.90),
                                           c(0.30 ,0.40 ,0.50 ,0.60 ,0.70 ,0.80),
                                           c(0.30 ,0.35 ,0.40 ,0.60 ,0.75 ,0.90),
                                           c(0.20 ,0.30 ,0.40 ,0.50 ,0.60 ,0.70),
                                           c(0.20 ,0.30 ,0.50 ,0.65 ,0.80 ,0.90),
                                           c(0.20 ,0.25 ,0.30 ,0.40 ,0.60 ,0.75),
                                           c(0.10 ,0.20 ,0.30 ,0.40 ,0.50 ,0.60),
                                           c(0.10 ,0.15 ,0.25 ,0.30 ,0.50 ,0.75),
                                           c(0.05 ,0.10 ,0.15 ,0.20 ,0.25 ,0.30),
                                           c(0.05 ,0.10 ,0.20 ,0.30 ,0.50 ,0.65),
                                           c(0.025,0.05 ,0.10 ,0.20 ,0.30 ,0.40)), 
   phi=0.3, phi1=0.6*0.3, phi2=1.4*0.3, maxtox=0.95, N=18, 
   cohortsize=3, maxN=9, maxNdec="STAY",acc_tit=0, design="BOIN", MTD_safer=TRUE)
}
\references{
Liu S, Yuan Y. Bayesian Optimal Interval Designs for Phase I Clinical Trials. 
    J R Stat Soc Ser C Appl Stat. 2015;64:507–23
    Yan F, Mandrekar SJ, Yuan Y. Keyboard: A Novel Bayesian Toxicity Probability Interval Design for Phase I Clinical Trials. Clin Cancer Res. 2017; 23: 3994–4003
}
