% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scenarios.R
\name{plot_scenarios}
\alias{plot_scenarios}
\title{Plot scenarios of true DLT rates}
\usage{
plot_scenarios(scenarios, save_as)
}
\arguments{
\item{scenarios}{scenarios of true DLT rates for each dose level. This also defines the number of dose levels}

\item{save_as}{where to save output}
}
\value{
plot
}
\description{
Plot scenarios of true DLT rates
}
\examples{
plot_scenarios(scenarios=list(c(0.40 ,0.50 ,0.60 ,0.70 ,0.80 ,0.90),
                              c(0.30 ,0.40 ,0.50 ,0.60 ,0.70 ,0.80)),
                              save_as="C:/Users/kdhollander/Desktop/test")
}
